Removed gluLookAt and Vector class.
The old vector class I had in this was pretty inefficient so I decided to
just scrap it and do it all by hand.

As a result I could also get rid of the gluLookAt / glGetFloatfv calls and
just construct the matrix myself.
Techincally the matrix isn't perfect right now which is seen by the fact
that I had to switch the signs on the movement input keys on some
directions.

Planning on putting this up on github soon so I'll be redoing the readme
next commit.